requirements.txt
streamlit
python-dotenv
pymupdf
python-docx
tqdm
numpy
pandas
faiss-cpu
sentence-transformers
